Move Attribute to shorter Namespace
→ \Ebln\Attrib\ForceFactory
