+++ tests/dataAttrib/code/EmptyProduct.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Test\Ebln\PHPStan\EnforceFactory\dataAttrib\code;
 
-use Ebln\PHPStan\EnforceFactory\ForceFactory;
+use Ebln\Attrib\ForceFactory;
 
 #[ForceFactory()]
 class EmptyProduct
