+++ tests/dataAttrib/code/ForcedFactoryProduct.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Test\Ebln\PHPStan\EnforceFactory\dataAttrib\code;
 
-use Ebln\PHPStan\EnforceFactory\ForceFactory;
+use Ebln\Attrib\ForceFactory;
 use Test\Ebln\PHPStan\EnforceFactory\dataAttrib\ForcedFactory;
 use Test\Ebln\PHPStan\EnforceFactory\dataAttrib\TraitFactory;
