+++ tests/dataAttrib/code/IndependentForcedFactoryProduct.php
@@ -6,7 +6,7 @@ namespace Test\Ebln\PHPStan\EnforceFactory\dataAttrib\code;
 
 use Test\Ebln\PHPStan\EnforceFactory\dataAttrib\IndependentFactory;
 
-#[\Ebln\PHPStan\EnforceFactory\ForceFactory(IndependentFactory::class)]
+#[\Ebln\Attrib\ForceFactory(IndependentFactory::class)]
 #[\INVALID\NOT\FOUND\ATTRIBUTE(IndependentFactory::class)]
 class IndependentForcedFactoryProduct
 {
