+++ tests/dataMixed/code/AttributeProduct.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Test\Ebln\PHPStan\EnforceFactory\dataMixed\code;
 
-use Ebln\PHPStan\EnforceFactory\ForceFactory;
+use Ebln\Attrib\ForceFactory;
 use Test\Ebln\PHPStan\EnforceFactory\dataMixed\AttributeFactory;
 
 #[ForceFactory(AttributeFactory::class)]
