+++ tests/dataMixed/code/MixedProduct.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Test\Ebln\PHPStan\EnforceFactory\dataMixed\code;
 
-use Ebln\PHPStan\EnforceFactory\ForceFactory;
+use Ebln\Attrib\ForceFactory;
 use Ebln\PHPStan\EnforceFactory\ForceFactoryInterface;
 use Test\Ebln\PHPStan\EnforceFactory\dataMixed\AttributeFactory;
 use Test\Ebln\PHPStan\EnforceFactory\dataMixed\ForcedFactory;
